% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsm.panel.R
\name{Rsm.panel}
\alias{Rsm.panel}
\title{Interactive panel to visualize and develop a response surface analysis}
\usage{
Rsm.panel()
}
\value{
A shiny panel with the response surface analysis, based on
the specification of the dependent and independent variables of
the data set provided, also provides various first- and second-order models,
both with and without interaction terms. Additionally, furnishes insights into
the optimal region of the surface and its corresponding location.
}
\description{
Interactive panel to visualize and develop a response surface analysis
}
\examples{
# Execute the function Rsm.panel()
# and provide the dataset. That's all!

}
